+++ src/Purview/Purviewdata.Autorest/custom/autogen-model-cmdlets/New-AzPurviewAzureKeyVaultObject.ps1
@@ -21,12 +21,13 @@ Create an in-memory object for AzureKeyVault.
 Create an in-memory object for AzureKeyVault.
 
 .Outputs
-Microsoft.Azure.PowerShell.Cmdlets.Purviewdata.Models.Api20211001Preview.AzureKeyVault
+Microsoft.Azure.PowerShell.Cmdlets.Purviewdata.Models.AzureKeyVault
 .Link
-https://learn.microsoft.com/powershell/module/Az.Purview/new-AzPurviewAzureKeyVaultObject
+https://learn.microsoft.com/powershell/module/Az.Purview/new-azpurviewazurekeyvaultobject
 #>
 function New-AzPurviewAzureKeyVaultObject {
-    [OutputType('Microsoft.Azure.PowerShell.Cmdlets.Purviewdata.Models.Api20211001Preview.AzureKeyVault')]
+    [Microsoft.Azure.PowerShell.Cmdlets.Purviewdata.ModelCmdletAttribute()]
+    [OutputType('Microsoft.Azure.PowerShell.Cmdlets.Purviewdata.Models.AzureKeyVault')]
     [CmdletBinding(PositionalBinding=$false)]
     Param(
 
@@ -39,7 +40,7 @@ function New-AzPurviewAzureKeyVaultObject {
     )
 
     process {
-        $Object = [Microsoft.Azure.PowerShell.Cmdlets.Purviewdata.Models.Api20211001Preview.AzureKeyVault]::New()
+        $Object = [Microsoft.Azure.PowerShell.Cmdlets.Purviewdata.Models.AzureKeyVault]::New()
 
         if ($PSBoundParameters.ContainsKey('BaseUrl')) {
             $Object.BaseUrl = $BaseUrl