+++ src/RecoveryServices/RecoveryServices.SiteRecovery.Test/ScenarioTests/E2A/AsrE2ATests.ps1
@@ -234,7 +234,7 @@ function Test-CreatePCMap
 	$Job = New-AzRecoveryServicesAsrProtectionContainerMapping -Name $ProtectionContainerMappingName -Policy $Policy -PrimaryProtectionContainer $PrimaryProtectionContainer 
 	WaitForJobCompletion -JobId $Job.Name
 
-	# Get protection conatiner mapping
+	# Get protection container mapping
 	$ProtectionContainerMapping = Get-AzRecoveryServicesAsrProtectionContainerMapping -Name $ProtectionContainerMappingName -ProtectionContainer $PrimaryProtectionContainer
 	Assert-NotNull($ProtectionContainerMapping)
 }
@@ -460,10 +460,10 @@ function Test-RemovePCMap
 	# Get the primary container
 	$PrimaryProtectionContainer = Get-AzRecoveryServicesAsrFabric -FriendlyName $PrimaryFabricName| Get-AzRecoveryServicesAsrProtectionContainer | where { $_.FriendlyName -eq $PrimaryProtectionContainerName }
 
-	# Get protection conatiner mapping
+	# Get protection container mapping
 	$ProtectionContainerMapping = Get-AzRecoveryServicesAsrProtectionContainerMapping -Name $ProtectionContainerMappingName -ProtectionContainer $PrimaryProtectionContainer
 
-	# Remove protection conatiner mapping
+	# Remove protection container mapping
 	$Job = Remove-AzRecoveryServicesAsrProtectionContainerMapping -ProtectionContainerMapping $ProtectionContainerMapping
 	#WaitForJobCompletion -JobId $Job.Name
 }
@@ -712,7 +712,7 @@ function Test-SiteRecoveryNewModelE2ETest
 	Assert-NotNull($Policy)
 	Assert-NotNull($Policy.Name)
 
-	# Get conatiners
+	# Get containers
 	$PrimaryProtectionContainer = Get-AzRecoveryServicesAsrFabric | Get-AzRecoveryServicesAsrProtectionContainer | where { $_.FriendlyName -eq $PrimaryProtectionContainerName }
 	Assert-NotNull($PrimaryProtectionContainer)
 	Assert-NotNull($PrimaryProtectionContainer.Name)
@@ -720,7 +720,7 @@ function Test-SiteRecoveryNewModelE2ETest
 	Assert-NotNull($RecoveryProtectionContainer)
 	Assert-NotNull($RecoveryProtectionContainer.Name)
 
-	# Create new Conatiner mapping 
+	# Create new Container mapping 
 	$Job = New-AzRecoveryServicesAsrProtectionContainerMapping -Name $ProtectionContainerMappingName -Policy $Policy -PrimaryProtectionContainer $PrimaryProtectionContainer -RecoveryProtectionContainer $RecoveryProtectionContainer
 	#WaitForJobCompletion -JobId $Job.Name
 
@@ -766,7 +766,7 @@ function Test-SiteRecoveryNewModelE2ETest
 	$Job = Remove-AzRecoveryServicesAsrNetworkMapping -NetworkMapping $NetworkMapping
 	#WaitForJobCompletion -JobId $Job.Name
 
-	# Remove conatiner mapping
+	# Remove container mapping
 	$Job = Remove-AzRecoveryServicesAsrProtectionContainerMapping -ProtectionContainerMapping $ProtectionContainerMapping
 	#WaitForJobCompletion -JobId $Job.Name
 	$ProtectionContainerMapping = Get-AzRecoveryServicesAsrProtectionContainerMapping -ProtectionContainer $PrimaryProtectionContainer | Where-Object {$_.Name -eq $ProtectionContainerMappingName}