+++ src/TrustedSigning/TrustedSigning/ChangeLog.md
@@ -19,8 +19,11 @@
 -->
 ## Upcoming Release
 
+## Version 0.1.2
+* Updated InvokeCIPolicySigning to support signing files with the .cip extension
+
 ## Version 0.1.1
-* Modified InvokeCiPolicySigning to include ShouldProcess command confirmation
+* Modified InvokeCIPolicySigning to include ShouldProcess command confirmation
 
 ## Version 0.1.0
 * Renamed from Az.CodeSigning